cepv-desktop/requirements.txt
PyQt5>=5.15.0
requests>=2.28.0
python-dotenv>=1.0.0

